Rearranged package infrastructre, no longer a separate utils directory, instead there is a utils sub-module and an errors sub-module
