+++ lightpath/errors.py
