+++ lightpath/utils.py
@@ -1,17 +1,20 @@
 ####################
 # Standard Library #
 ####################
+from enum import Enum
 
 ####################
 #    Third Party   #
 ####################
 from ophyd import Device, EpicsSignalRO, Component
 from ophyd.utils.epics_pvs import raise_if_disconnected, AlarmSeverity
+from super_state_machine.machines import StateMachine
+from super_state_machine.extras   import PropertyMachine
+from super_state_machine.errors   import TransitionError
 
 ####################
 #     Package      #
 ####################
-
 
 class MPS(Device):
     """
@@ -76,4 +79,36 @@ class MPS(Device):
         """
         self._run_subs(sub_type = self.SUB_MPS,
                        faulted = self.faulted) 
+
+class DeviceStateMachine(StateMachine):
+    """
+    State Machine for device transitions
+    """
+    class States(Enum):
+        """
+        Four way state
+        """
+        INSERTED   = 'inserted'
+        REMOVED    = 'removed'
+        UNKNOWN    = 'unknown'
+
+
+    class Meta:
+        allow_empty   = False
+        initial_state = 'unknown'
+
+
+class LoggingPropertyMachine(PropertyMachine):
+    """
+    Creates a property in the parent device that uses the built in logger
+    """
+    def __init__(self, machine_type):
+        super().__init__(machine_type)
+
+    def __set__(self, obj, value):
+        old_value = self.__get__(obj)
+        super().__set__(obj, value)
+        value = self.__get__(obj)
+        obj.log.info('Change state on %r from %r -> %r',
+                     obj, old_value, value)
