--- lightpath/utils/__init__.py
@@ -1,2 +1,0 @@
-from .errors import *
-from .utils  import DeviceStateMachine, LoggingPropertyMachine
